# Conflicts:
#	.idea/workspace.xml
#	target/classes/com/yulan/pojo/ItemMLGY.class
#	target/classes/com/yulan/service/impl/ItemServiceImpl.class
#	target/classes/com/yulan/utils/FileUpload.class
#	target/yulan-order.war
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/Ctm_orderController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/CustomerBalanceController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/ExceptionResolver.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/InfoStateController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/ItemController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/PackDetailController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/controller/YLcontractentryController.class
#	target/yulan-order/WEB-INF/classes/com/yulan/encode/CommodityEncode.class
#	target/yulan-order/WEB-INF/classes/com/yulan/encode/CurtainCartItemEncode.class
#	target/yulan-order/WEB-INF/classes/com/yulan/encode/CurtainCommodityEncode.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/AreaCode.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/AreaDistrict.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/AreaRegion.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/AreaUser.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/BrandType.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Ctm_order.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/CurtainOrder.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Customer.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/CustomerBalancePeriod.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/CustomerBalancePeriodDetail.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/CustomerType.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/HesEmployee.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Item.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/ItemMLGY.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Pack_detail.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/SalPromotion.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/SalReturnRecord.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Sal_promotion.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/Sal_rebate_certificate.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/YLcontract_v2015.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/YLcontract_v2015_paa.class
#	target/yulan-order/WEB-INF/classes/com/yulan/pojo/YLcontractentry.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/ActivityGroupTypeServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/Area_ownerServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CartItemServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CommodityOrderServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CurtainCartItemServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CustomerAssignmentsServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CustomerBalanceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/CustomerinfocardgroupServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/InfoStateServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/SalPromotionServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/Web_userServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/service/impl/webcontractwfmemberServiceImpl.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/BackUtil.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/BlueVegetableBean.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/FileUpload.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/MD5.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/MapUtils.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/ObjectUtils.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/StringUtil.class
#	target/yulan-order/WEB-INF/classes/com/yulan/utils/TransferXML.class
#	target/yulan/WEB-INF/classes/com/yulan/service/impl/ItemServiceImpl.class

+++ src/main/java/com/yulan/pojo/ItemMLGY.java
@@ -9,7 +9,7 @@ public class ItemMLGY {
 
     private String itemType;
 
-    private Short no;
+    private Double no;
 
     private String parentItemNo;
 
@@ -52,11 +52,11 @@ public class ItemMLGY {
         this.itemType = itemType == null ? null : itemType.trim();
     }
 
-    public Short getNo() {
+    public Double getNo() {
         return no;
     }
 
-    public void setNo(Short no) {
+    public void setNo(Double no) {
         this.no = no;
     }
